refactor: integrate UserInfo and OAuthUserInfo features
