+++ src/main/java/toy/blog/be/controller/follow/dto/FollowResponse.java
@@ -3,7 +3,7 @@ package toy.blog.be.controller.follow.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import toy.blog.be.domain.entity.OAuthUserInfo;
+import toy.blog.be.domain.entity.UserInfo;
 
 import java.util.List;
 
@@ -11,5 +11,5 @@ import java.util.List;
 @Builder
 @AllArgsConstructor
 public class FollowResponse {
-    List<OAuthUserInfo> oAuthUserInfos;
+    List<UserInfo> userInfos;
 }
