--- src/main/java/toy/blog/be/domain/entity/OAuthUserInfo.java
@@ -1,52 +1,0 @@
-package toy.blog.be.domain.entity;
-
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicUpdate;
-import toy.blog.be.infra.IdGenerator;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@DynamicUpdate
-@Entity
-@Getter
-public class OAuthUserInfo {
-    @Id
-    private String id;
-
-    private String name;
-    private String email;
-    private String provider;
-    private String nickname;
-    private LocalDateTime createAt;
-    private LocalDateTime modifiedAt;
-
-    @Builder
-    public OAuthUserInfo(String name, String email, String provider, String nickname) {
-        this.id = IdGenerator.newId();
-        this.name = name;
-        this.email = email;
-        this.provider = provider;
-        this.nickname = nickname;
-        var now = LocalDateTime.now();
-        this.createAt = now;
-        this.modifiedAt = now;
-    }
-
-    public OAuthUserInfo update(String name, String email) {
-        this.name = name;
-        this.email = email;
-        this.modifiedAt = LocalDateTime.now();
-        return this;
-    }
-
-    public void updateNickname(String nickname) {
-        this.nickname = nickname;
-        this.modifiedAt = LocalDateTime.now();
-    }
-}
