+++ src/main/java/toy/blog/be/domain/value/UserProfile.java
@@ -2,7 +2,7 @@ package toy.blog.be.domain.value;
 
 import lombok.Getter;
 import lombok.Setter;
-import toy.blog.be.domain.entity.OAuthUserInfo;
+import toy.blog.be.domain.entity.UserInfo;
 
 @Getter
 @Setter
@@ -12,8 +12,8 @@ public class UserProfile {
     private String provider;
     private String nickname;
 
-    public OAuthUserInfo toOAuthUserInfo() {
-        return OAuthUserInfo.builder()
+    public UserInfo toOAuthUserInfo() {
+        return UserInfo.builder()
                             .name(name)
                             .email(email)
                             .provider(provider)
