+++ src/main/java/toy/blog/be/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package toy.blog.be.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import org.springframework.stereotype.Repository;
 import toy.blog.be.domain.entity.UserInfo;
 
-@Repository
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<UserInfo, String> {
+    Optional<UserInfo> findByEmailAndProvider(String email, String provider);
+    boolean existsByNickname(String nickname);
 }
